refactor: remove block committer code
