+++ crates/starknet-committer/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "starknet-committer"
+version.workspace = true
+edition.workspace = true
+repository.workspace = true
+license.workspace = true
+license-file.workspace = true
+
+[dependencies]
+committer = { path = "../committer", features = ["testing"] }
+hex.workspace = true
+log.workspace = true
+rstest.workspace = true
+serde_json.workspace = true
+starknet-types-core.workspace = true
+thiserror.workspace = true
+tokio.workspace = true
+
+[lints]
+workspace = true
