+++ crates/starknet-committer/src/block_committer/errors.rs
@@ -1,6 +1,6 @@
 use thiserror::Error;
 
-use crate::forest_errors::ForestError;
+use crate::starknet_forest::forest_errors::ForestError;
 
 #[derive(Debug, Error)]
 pub enum BlockCommitmentError {
