+++ crates/starknet-committer/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod block_committer;
+
